requirements.txt
weaviate-client[agents]
python-dotenv
rich

